+++ shared/GR_Dataset.lua
@@ -63,6 +63,7 @@ function ds:WhatsNew() -- What's new in the current version
             - Can see anti-spam list, but not change it.
             - Reworked black list and added a privacy option for reason.
             - Reworked invalid zones so you can specify a name of a zone to ignore.
+        - Made races by faction again.
     |CFFFFFF00v3.1.40 Notes|r
         - Added option to keep addon running and ignore certain ways to close it.
         - esMX (Spanish Mexico) localization added.
@@ -244,9 +245,10 @@ function ds:races() -- Race data
             [37] = "Alliance",   -- Mechagnome
         }
 
+        local myFaction = UnitFactionGroup('player')
         local raceInfo = C_CreatureInfo.GetRaceInfo(raceID)
         if raceInfo then
-            if raceFactions[raceInfo.raceID] then
+            if raceFactions[raceInfo.raceID] == myFaction then
                 tbl[raceInfo.raceName] = {
                     ['id'] = raceInfo.raceID,
                     ['name'] = raceInfo.raceName,